Split mitigation api into new python file
