+++ app/routes/api.py
@@ -1,6 +1,4 @@
 import logging
-import json
-import os
 from app.models import CoOccurrence, Platform, db, Tactic, Technique, Mismapping, AttackVersion, DataSource
 from app.models import technique_platform_map, tactic_technique_map, tactic_ds_map, technique_ds_map
 from app.routes.auth import disabled_in_kiosk
@@ -24,8 +22,6 @@ from sqlalchemy import asc, func, distinct, and_, or_, literal_column
 from sqlalchemy.dialects.postgresql import array
 from sqlalchemy.orm.util import aliased
 
-from app.domain.mitigations_service import MitigationsService
-
 logger = logging.getLogger(__name__)
 
 api_ = Blueprint("api_", __name__)
@@ -46,19 +42,6 @@ def get_versions():
 
     return jsonify(version_strs), 200
 
-@api_.route("/api/mappings", methods=["GET"])
-@wrap_exceptions_as(ErrorDuringAJAXRoute)
-def get_mappings():
-    """Returns a list of custom mapping data stored in JSON file (JSON response)"""
-    g.route_title = "Get Custom Mappings"
-    technique = request.args.get("technique")
-    if (technique is None):
-        logger.error("request failed - technique field missing")
-        return jsonify(message="'technique' field missing"), 400
-    logger.info("querying custom mappings")
-    ms = MitigationsService()
-    return_value = ms.get_mitigations(technique)
-    return jsonify(return_value), 200
 
 @api_.route("/api/mismappings", methods=["GET"])
 @wrap_exceptions_as(ErrorDuringAJAXRoute)
